requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
camelot-py[cv]<0.11.0
tabula-py==2.8.2
pytesseract==0.3.10
PyMuPDF==1.23.8
Pillow==10.1.0

# AI/ML
google-generativeai==0.3.1


sentence-transformers>=2.6.0
huggingface_hub>=0.20.0

faiss-cpu==1.7.4
numpy==1.24.3

# Additional dependencies
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

